fix(app): Prevent crash when ASN database is missing
The application would panic and exit on startup if the ASN TSV database
file specified in the configuration was not found on disk. This was caused
by a regression where a check was added that unconditionally required the
file to exist.

This commit changes the behavior to be more resilient. If the ASN
database file is missing or fails to load, the application will now log a
warning and continue to run with the enrichment feature disabled, instead
of crashing.

This improves the user experience by not requiring a dummy file to be
created for the application to run.